Updated user dropdown to properly use React-Redux.
